// apps/expo-meal/src/types/types.d.ts
export type Category = {
	id: string;
	title: string;
	color: string;
};

export type Meal = {
	id: string;
	categoryIds: string[];
	title: string;
	affordability: string;
	complexity: string;
	imageUrl: string;
	duration: number;
	ingredients: string[];
	steps: string[];
	isGlutenFree: boolean;
	isVegan: boolean;
	isVegetarian: boolean;
	isLactoseFree: boolean;
};